debug

+++ frontend/src/main/scala/com/neo/sk/tank/front/tankClient/BrickClientImpl.scala
@@ -46,8 +46,8 @@ object BrickClientImpl{
 //    ctx.fillText(curBlood.toString,(position.x + offset.x - model.ObstacleParameters.halfBorder) * canvasUnit,(position.y + offset.y - model.ObstacleParameters.halfBorder) * canvasUnit,14)
     ctx.fillStyle = "#8B6969"
     ctx.fillRect((position.x - model.ObstacleParameters.halfBorder + offset.x) * canvasUnit
-      ,(position.y + model.ObstacleParameters.halfBorder + offset.y - model.ObstacleParameters.border * curBlood/ model.ObstacleParameters.BrickDropBoxParameters.blood) * canvasUnit,
-      model.ObstacleParameters.border * canvasUnit,(model.ObstacleParameters.border * curBlood/ model.ObstacleParameters.BrickDropBoxParameters.blood) * canvasUnit)
+      ,(position.y + model.ObstacleParameters.halfBorder + offset.y - (model.ObstacleParameters.border * curBlood).toFloat/ model.ObstacleParameters.BrickDropBoxParameters.blood) * canvasUnit,
+      model.ObstacleParameters.border * canvasUnit,((model.ObstacleParameters.border * curBlood).toFloat/ model.ObstacleParameters.BrickDropBoxParameters.blood) * canvasUnit)
     ctx.strokeStyle = "#8B6969"
     ctx.lineWidth = 2
     ctx.beginPath()
